debt: auth.js: extract invite-flow to ./auth/invite-flow.js
Invitation creation + QR rendering + TTL countdown lift out of
auth.js into a small sibling module. qrcode import moves with it.
inviteCountdown is now module-local in invite-flow.

auth.js: 728 → 575 lines.

Also drops the unused clearOverrideTtlTimer export from
ws-command-handlers.js that knip flagged.

npm test:unit 782/782 pass. knip clean.
