+++ server/lib/ws-command-handlers.js
@@ -195,7 +195,4 @@ function clearOverrideTtlTimer() {
   }
 }
 
-module.exports = {
-  handleWsCommand,
-  clearOverrideTtlTimer,
-};
+module.exports = { handleWsCommand };
